Added Tested Modules Folder and Basic Detection Code
In the folder "Tested Modules" stand alone versions of modules shall be uploaded. Those modules may or may not be included in our final project.

Basic Detection Test is code. It uses a sliding window to compare a template with an input image and returns a score-map of the comparison